Issue template update
The issue template is used by GitHub when a new issue is created.
Currently its main purpose is to prompt the issue reporter in to
providing all the information required to triage and investigate the
issue.

The main purpose of this patch is to give the individual working on
fixing the issue better guidence as to the steps that need to be taken
to complete it. This includes requiring the the individual to keep the
description (the section populated by the template) up-to-date with the
progress on work that has been done, e.g. pull requests created or
decisions made. In addition this patch seeks to make a clearer
distinction between the responsibilities of the issue reporter and the
issue assignee.