Added data disk size precheck
In order to solve an issue where the deploy would fail if a small
secondary disk exists on the system. This is becuse the deploy.sh
script would try to use the small disk as a data partition. Added
a precheck within deploy.sh to ensure any potential data devices
are large enough. The minimum disk size is pulled from the
main.yml playbook in OSA

Connects #1235