Enumerate log file targets and multiline config
Using more than one prospector allows us to configure the openstack
services with multiline log parsing. This prevents us from applying
that same logic to non-OpenStack services (which will mess up log
shipping). It also allows us to define multiline rules for each of
the services we hope to ship to logstash with multiline log messages.

Filebeat will ignore directories/paths that do not actually exist on
the host so we can safely enumerate all of the services we want to
apply multiline parsing to in one prospector.

We *do* have to, however, ensure that paths do not overlap with the
prospectors. Filebeat explicitly calls this out as "undefined"
behaviour.

Included near the multline configurations are play.golang.org links so
others can experiment with what we used to test the multiline regular
expressions. Also, given the complexity and widespread use of our
OpenStack multiline expression, we've attempted to thoroughly document
it.

Connects to rcbops/u-suk-dev#261