Update Contributing guidelines
Swim lanes allow the board to be split into categories of work that
represent the different work streams the team needs to manage. This
commit updates the contributing guidelines to define the available swim
lanes and how to use them to manage WIP.

This change updates the contributing guidelines to further clarify what
constitutes a single item of WIP.

The wording around blocking WIP has been adjusted to highlight the
importance of removing the blockage quickly.

The details regarding where to find the current WIP limits have been
emended and the use of a minimum WIP limit has been explained.

The role of support in identifying new WIP has been made explicit.

Document the use of the label status-dont-merge. Currently the
documentation only provides use of the label status-blocked which is not
appropriate when the pull request is simply being revised.

Document the use of status-abandoned, this label allows filtering based
on WIP that is closed without being completed.

In general an upstream fix is not immediately consumable by RPCO, OSA
has processes to periodically update its SHAs to pull in newer code from
associated projects. This lag between a fix appearing in a code base and
being available to RPCO has the potential to block new work. This commit
adds documentation for a new column, Pending-SHA-update, that can be
used as a buffer so that this work is appropriately tracked but does not
block new WIP. In addition a new label, status-pending-sha-update, is
added to enable the appropriate Waffle integration.

Connected https://github.com/rcbops/u-suk-dev/issues/268