Use proper entity name in verify_maas.yml
verify_maas.yml uses the entity name in a call to the maas
tool to verify the configuration of maas checks. Prior
to this commit verify_mass.yml used the inventory_hostname
fact to set the entity name it would check. However, the
host_setup.yml rpc_mass task would use a concatination of
inventory_hostname and the maas_fqdn_extension facts to
set the entity name on creatiion. This means that if
mass_fqdn_extension is non-empty the verify maas playbook
would fail.

This commit fixes this issue by changing the entity name
variable in verify_maas.yml to match how it is set up in
the host_setup.yml task.

Connects #941