Install virtualenv pip package for rpc-maas role
In order to install the maas environment inside a virtualenv, we need to
ensure we are actually installing the virtualenv pip package. In an AIO
this is done by the nova role on the compute 'container', meaning it
actually gets installed on the host, but in a multinode we need to do
this explicitly.

Connected https://github.com/rcbops/rpc-openstack/issues/1301