use correct location for maas plugins
With the switch to using maas from a venv, we moved to using the
maas_plugin_dir var as a reference for the location of the installed
plugins.  The value of this var, however, was the location of plugins
inside the git clone of rpc-openstack, rather than the location of
where the plugins get installed. This meant that only a deployment host
had the plugins in the correct location.

This commit switches the value of this var to point to the correct
installed location, whilst creating a new var to be used by the only
other task that was using the original value. It is done in this way to
minimise the amount of change needed to fix this issue.

Additionally, some areas where we were hardcoding the path to the
plugin location are now set to use the var value.

Connected https://github.com/rcbops/rpc-openstack/issues/1303