Enable active-active balancing of repo servers
Prior to this commit repo servers were configured in an F5 pool
which assigns a unique priority to each server. This means
that all requests are forwarded to the first server in the pool
until that server become unavailible. This is active-passive
load balancing.

This commit disables prioritization of of repo servers, which
enables active-active load balancing and forwards connections
under the fastest-connection policy. All servers in the pool
are now treated equally. This change is benificial in that the
load is now spread throughout the repo servers and the
deployment can benifit from the combined power.

Connects #190