Ensure filebeat (re)starts only once on install
The filebeat playbooks were causing a start followed very quickly by a
restart of the filebeat service, which was causing the service to die.

This commit adds a 'flush_handlers' so that the service is (re)started
prior to the task where it is 'enabled' thereby removing the need for a
second restart.  This fixes the issue with the filebeat service
stopping.

Connected https://github.com/rcbops/rpc-openstack/issues/1310

Co-Authored-By: Ian Cordasco <ian.cordasco@rackspace.com>