Add OSA upgrade script to test-upgrade.sh
One of the steps when performing an upgrade from Liberty to Mitaka will
be to run OSA's upgrade script, run-upgrade.sh. To enable us to perform
a tested upgrade this commit adds run-upgrade.sh to the RPCO test script
test-upgrade.sh.

A release note is also added to ensure that security hardening is
disabled where necessary. Currently deploy.sh manages that with an
enviroment variable but that is not feasible in this situation.

Connected https://github.com/rcbops/u-suk-dev/issues/392