Resolve pip_lock_down and maas_common issues
The rpc_maas role needs to depend on the pip_lock_down role, otherwise
the openstack clients and other pip packages are not installed from the
repo server.

To prevent this issue occuring on other roles that perform pip installs,
and to keep the roles consistent, this PR adds the pip_lock_down
dependency to the rpc_support and horizon_extensions roles.

This will prevent us from having to run "pip-lockdown" against hosts
unnecessarily.

Additionally, in the next version of python-keystoneclient
apiclient.exceptions will be removed, as per:
https://review.openstack.org/#/c/257127/ - we can fix this in advance by
following the recommendation and use keystoneclient.exceptions. This
won't affect current releases but will future proof for upgrades to the
Newton cycle (and above).

Connects https://github.com/rcbops/rpc-openstack/issues/1404