Update release notes based on r13.0.0rc2 review
The release notes for r13.0.0rc2 were reviewed in preparation for the
release of r13.0.0. Of the changes to be agreed based on that review the
following are relevant to newer versions of the notes:

- a test note in the repo is removed
- a note about L3HA has been expanded to explain why it's being disabled
- the prelude sections have been removed because they weren't adding
  value
- a note regarding overlay networks has been reworded to better explain
  the change to which it refers
- security hardening note has been adjusted to keep to control interface
  consistent for all deployment paths
- the removal of LVMISCSIDriver by Cinder is noted

In addition to the changes in the cherry-pick, the upgrades sections of
the notes have been removed because upgrades are not supported in
r13.0.0.

Connected https://github.com/rcbops/u-suk-dev/issues/398

(cherry picked from commit 9d64ec35d54beea889b89d2dd5ae92d797d788f5)