Move beaver restart to it's own play
The way we currently do the beaver restarts means that it takes an
inordinate amount of time to restart beaver across all hosts.

This commit moves the beaver restart tasks to their own play so that
they can be targeted at the correct hosts in the way that ansible is
designed to be used.

Connected #1419