Ceph hostname change during upgrade
L>M has changed the hostname of our containers, therefore we
need to adapt this everywhere. Ceph was left behind, and this
is fixing the hostnames by shutting down the monitors, removing the mons,
deleting the containers, recreating them, and rejoining the cluster, all
of this in a serial way.

The ceph-mon playbook was adapted to verify the bind mount is there, in
order to avoid data loss.

Comments are added explaining what is done, and what is needed
when upgrading an environment with ceph.

This commit also ensures the generation of the facts used in the
setup-maas is properly done (explicitly asking the generation of
the facts by including our new gen-facts playbook).

Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>
Connected rcbops/u-suk-dev#383

(cherry picked from commit acb2491e89e163c78fe776524f31925bd008de54)