Check swift for bad rings
These changes create a swift object within a container which is checked
by a MaaS check (lb_api_check_swift_access) to confirm swift is
operating properly. If a return code other than 200 is received, swift
is not returning object data. Generally, a 500 indicates the swift ring
is corrupt in some way.

Support may want to take action given the result of this check to remove
swift nodes from the ring, but we do not believe this process should be
automated without confirmation.

Additionally, a load balancer check could be created on an F5 to verify
access to this file. Any proxy not able to access the file properly
could be removed, though again this may not be the most appropriate way
to resolve the issue.

This may not be the optimal place to create the object to check ring
status, given customers may want something similar without utilizing
MaaS for health checks, but as of now this is probably good enough.

Connects: #800