Search for and provide domain in horizon check
In multi-domain deployments of RPC OpenStack, Horizon will require the
user to enter a domain name. As such, we need to provide that
information when posting our login request to Horizon in our MaaS check
plugin.

To detect the multi-domain case, we search for an input field named
"domain" and only then do we provide it. To make the check slightly more
efficient we also refactor it to just parse the HTML once instead of 2
or 3 times.

Connects #1262