Add ELK upgrade support for L to M upgrades
ElasticSearch, Logstash, Kibana (ELK) provides a single dashboard from which to view the logs on a deployment. The versions of these applications in the RPCO r12 series is out of date and affected by CVEs that need to be addressed by moving to newer versions. This change updates the deployment to use the versions supported in our r13 series releases.

ElasticSearch is upgraded from 1.7 to 2.x, as part of this change the data has been migrated because the format in 1.7 is not compatible with 2.x.

The migrated ElasticSearch data cannot currently be viewed in Kibana.

Connected rcbops/u-suk-dev#237
(cherry picked from commit d14f01c9ec62562cb5cb5212f268286e4567fff6)