Check status of ceph cluster before rebuilding
This commit checks the status of the cluster before rebuilding on each
iteration of the loop to ensure that the cluster is healthy before
proceeding.  If this isn't done and attention isn't being paid to the
upgrade process then all containers could be torn down and data lost.

Connects https://github.com/rcbops/u-suk-dev/issues/471