Add release note regarding ElasticSearch upgrade
The ElasticSearch upgrade process, performed as part of a Liberty to
Mitaka upgrade, requires the Ansible var 'logging_upgrade' set to true.
Failure to do so may result in data corruption. This commit adds a
release note to document this issue for r13.1.0.

Connected https://github.com/rcbops/u-suk-dev/issues/550