Remove reindex-wapper tag from 2 tasks
The reindex-wrapper tag was incorrectly applied to 2 removal tasks.
The first task removes legacy plugins of which the reindex-liberty.py uses
to do the actual reindexing. The other removes the legacy apt packages
which is less of an issue but may be needed again during a pre-upgrade
reindexing.

Connects https://github.com/rcbops/u-suk-dev#1550