Remove duplicated ansible.cfg configurations
This patch removes duplicated Ansible configurations that match OSA's
defaults. The ssh timeout was left in the file since it differs from
the OSA default (120 seconds in RPC vs 5 seconds in OSA).

RPC previously set forks at 15 for all deployments, but OSA uses a
safer mechanism to determine the maximum amount of forks. It sets a
hard limit of 10 but it lowers that limit to the maximum amount of
CPUs in the system if the CPU count is under 10.

Connects rcbops/rpc-openstack#1639