Skip yesterday's backup check when deploy is new
Recently a MaaS check for holland was introduced which checks that
yesterday's backup exists. This makes sense for production but not for
gate runs. In the same PR a mechanism for excluding checks from local
MaaS verification was added, however not all gate runs use local verification
and no mechanism was included for excluding the check from full MaaS API
verification.

This commit attempts to guess the age of a machine by stat-ing
/etc/hostname.  If that file is less than a day old, then the check is
skipped with a successful status returned. This means that gate runs
will pass with no backups, and production deploys will be indicating
backup failures before they are handed over to customers.

Also:
 * Removes local verification check exclusion mechnaism as its no longer used.
 * Removes duplicate state keys in holland_preinstall

Connects rcbops/u-suk-dev#950