Set openstack_release to the value of rpc_release
To ensure that the repo build process, venvs, etc all make
use of the RPC version tag instead of the OSA version tag
this patch sets the openstack_release to the value of the
rpc_release variable.

Due to the structure of how RPC wraps OSA this requires the
rpc_release variable to be set in a user_*.yml override file.
This is not an ideal approach, but I can't see any other
way to do it in the current folder structure.