Use to_json filter to properly format unicode
For some reason in Newton the filebeat paths and tags are being
populated into the jinja2 template as unicode strings.  The leading
u is causing filebeat to treat the path as not found, hence it is
not sending any logs.  Using the to_json filter causes jinja2 to
properly format the string allowing the path to be read and the
correct tag(s) to be applied.

Connects #863

(cherry picked from commit a9f21bca9e050ef4990c4b2e382dde46d610e542)