[backport] HORIZON_SSL now uses HTTP monitor
And change what monitor the HORIZON_SSL pool is using.
This commit configures the HORIZON_SSL pool to use the
HTTP monitor which checks the backend nodes using port 80.
This commit fulfills the feedback given in:
https://github.com/rcbops/rpc-openstack/pull/1668#issuecomment-272914892.

The reason this is done is due to the new behavior introduced
in OSA Newton. SSL is expected to be terminated at the LB, and
is no longer implemented on the horizon containers. For more
information on that change, please see:
https://review.openstack.org/#/c/277199/

Connects https://github.com/rcbops/rpc-openstack/issues/1632

(cherry picked from commit b40cd5009779e7e2991058af8bd01c14fb16fcf7)