Update CONTRIBUTING
This commit attempts to slim down the CONTRIBUTING guidelines, leaving only
the steps necessary for someone to contribute a fix for a bug, add documentation,
and/or enabling an upstream service in RPCO.

The goal here is to simplify the document so that things specific to
the dev team's process do not sufficate the content that is specific
to someone contributing a bug fix or a new feature.

It's important to note that none of the content will be lost.
All of the content that has been stripped has either 1) been moved to
RELEASING.md or 2) is repeated in the workflow document here:
https://github.com/rcbops/u-suk-dev/blob/master/WORKFLOW.md

Connects https://github.com/rcbops/u-suk-dev/issues/833