search all pids and children for process checks
cgroup.procs only displays one layer, sometimes containers will have
child cgroups that we want to check.  This change allows us to switch
to a check that checks all descendants of a parent as well.

We do not use the lxc module here for two reasons.  One, it is built
only for python3, and two, it is not available in pypi (and is not
fun to build out of tree).

Connected: https://github.com/rcbops/u-suk-dev/issues/1001