create distro specific maas process check symlink
Xenial and Trusty both support different levels of cgroup functionality.
This has caused us to need different behavior for checking the status of
processes within containers.  In this commit we allow for both ways,
using a symlink to link to the correct version based on the Ubuntu
version installed.

Connects: https://github.com/rcbops/u-suk-dev/issues/1101