Remove ssl_impossible from console services
Traffic to this service is now expected to be SSL-terminated
at the load balancer, which means a public, ssl virtual server
is needed.
This commit removes ``ssl_impossible`` from the console service's
POOL_PARTS. This will tell the script to create a public ssl
virtual server for the services.

Connects https://github.com/rcbops/rpc-openstack/issues/1977