Bump version of reno to 2.2.0
We need to include some fixes that were made into the reno
code base in order for RPCO to start utilizing reno. This
commit bumps the version to the latest as of March 20th, 2017.

In addition, this commit changes the depth value when travis
clones the rpc-openstack repository. This is so reno can
properly traverse the commit history it needs to produce
release notes.

Connects https://github.com/rcbops/u-suk-dev/issues/1421