Add newton-14.0 to source directory
I hope I worded that right. Anywho, newton-14.0 has been added
into the source directory, so that sphinx will now build release
notes for the latest release, newton-14.0.

NOTE: The newton-14.0 reno release notes shouldn't be taken that
seriously, since we have not been diligent to create a release
file for most of the work done in newton-14.0. Reno releasenotes
should only be relied on for versions newton-14.1 and greater

An additional step is added into .travis.yml so that the
rpc-openstack clone is replaced by a full clone. This is
is reno can traverse all the needed refs in order for it to
function.

Connects https://github.com/rcbops/u-suk-dev/issues/216