Update test-requirements
- github3.exceptions doesn't exist

When simply installing from pip install github3.py, you'll get
the 0.9.6 version.
If no pinning is done, the github3 library would be an older
version that doesn't have github3.exceptions, needed for this
script.

- rpc-differ is not found
Release.py makes use of rpc-differ, calls it, but never
explicitly requires it.

- osa_differ 0.2.1 pip package doesn't ship templates
Latest version of rpc_differ doesn't specify the minimum
osa_differ version. You could end up with osa_differ 0.2.1
that doesn't ship the necessary templates for rpc. This
forces the version of osa_differ to 0.2.2.

- pandoc is not found
Pandoc is not listed in the requirements, but still used