Deploy ARA
Install the Ansible Run Analyser callback. This is used to gather task
failure information in gate jobs. A junit stream is created from failed
tasks, this is reported to jenkins so that failed tasks are reported as
failed unit tests. This allows a user to quickly see a list of failed
tasks rather than digging through logs.

Connects rcbops/u-suk-dev#318