Add pre-upgrade nova db Hygiene tasks
This change adds two tasks to run a Database deleted instances archive
and then to prune the shadow tables in preparation for upgrade. This
change will ensure that all deleted instances are cleanup which speeds
up nova migrations as well as ensures migrations are not impeded due
to legacy data being stuck in a deleted VM record.

Closes-Bug: #1608
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>