[kilo] Correct ansible-role-requirements
The openstack-ansible ansible-role-requirements.yml file
is specified in galaxy format and no longer works. In this
patch we ensure that the same role is specified using the
git format so that it's always retrievable. This ensures
that the ansible bootstrap works properly.

Also, the ceph roles which were specified in this file are
removed as this file is never used. The ceph roles are
instead included using submodules.