Remove old beaver checks/alarms (#2239)
Our gate job RPC-AIO_mitaka-upgrade-periodic, which does a more
extensive maas test, is failing verify-maas.yml after doing the mitaka
upgrade.  This is because on mitaka we remove beaver in favour of
filebeat but nothing cleans up the old beaver MaaS checks.

This commit simply removes those old checks as part of the beaver
cleanup.