Add git artifact tooling
The git artifacting is a protection against upstream
git repositories disappearing or being renamed.

In order to ensure that existing artifacts are not
removed during the mirror upload, the artifacts are
first pulled, then any configured repositories are
updated, then all the data is pushed back up again.

When updating the repositories, the 'master' branch
is pulled instead of the set SHA due to the fact that
the git repository on rpc-repo is used by multiple
series and the checked out SHA therefore does not
matter, as long as a git pull from there contains
the whole git repo.