Automatically clean-up lxc-images folder
When replacing the container artifacts for an
existing published release, the upload process
currently removes the old build ID from the
index, but leaves the old build data behind.

This patch implements a change to the way the
data is uploaded so that any old build data for
the same release version is cleaned up.