Use group_vars in RPC-O (#2245)
* Move to group_vars folder structure for rpc

* Move user variables to group_vars where possible

With overridable group_vars now allowed, we should move away
from using user_* variables, unless really necessary.

This would leave a product with default variables, but still
have a completely overridable group_vars in userspace.

* Cleanup more variables

These variables for ceph and maas were either unused,
or equal to defaults. No need to carry them around.

* Move rpc variables default to group vars

Last set of vars to be moved

* Carry a dummy file to pass gates

Valid until removal of the useless tasks in gates

* Add OSA default variables for RPC playbooks

RPC playbooks currently read their group vars, and the user
provided group vars.

This extend the group vars list with the ones provided to OSA.
Therefore if an RPC playbook need an OSA group var, it will
work out of the box.

* Don't use group vars for bootstrap aio

The bootstrap AIO should not make use of the vars plugins. It will
lockdown on the first pip_install, which is not welcomed.

This ensures the deploy.sh modifies the openstack-ansible.rc
to always include our var plugins when ready to do so.

* Override pkg_locations

Packages locations are not group vars overridable (defined
staticly in the playbook).

* Define ceph_stable

ceph_stable was removed during the cleanup because this var
is included in stable/ocata, but it's not included in
openstack-ansible stable/newton. We should keep this variable
until then.

* Updated documentation

Removed all references to user_rpco_variables_defaults, because
they aren't useful anymore.

* Removed variables defaults from bootstrap

Bootstrap AIO doesn't need to copy/update the /etc/openstack_deploy
/user_rpco_variables_defaults.yml anymore.

The upgrade jobs should take care of the removal of this file,
if any is present.

* Remove unused file

With all the previous changes, user_rpco_variables_default
doesn't need to exist anymore.

* Update release.py script with new vars

The string storing the release version
is now stored in ``group_vars/all/release.yml``.

* Add a release note

RPC-Openstack had multiple levels of variable precedence, and we
heavily relied on user variables to complement/define
OpenStack-Ansible variables. This is not required anymore.
RPC-Openstack can now ship safe defaults for OpenStack-Ansible
without needing to set user variables. The user variables can
still be used by the deployer, but we now use the group_vars/
and host_vars/ folder for safe defaults in tree.

RPC-O previously relied on user-space configuration files to
set variables used in playbooks across both the OpenStack-Ansible
and RPC-O playbooks. These user-space configuration files provided
global overrides only and could not be scoped to specific host
groups. These defaults have now been set in group_vars in the code
tree, reducing the need to overwrite user-space files, making them
easier to scope to host groups, and reducing their precedence level
to improve future development flexibility.

* Docs: Explain the user_osa_variables_defaults

In user_osa_variables_defaults we override openstack_release and
pkg_locations

This file explains why it's done there and not in group_vars/

* Avoid long discussions

Dropping these files to keep the world the way it used to work.
Change is bad.

* No need for a support file override

Because we keep the useless user_* files.

* Move variables around

We keep the old ideas of having the user variables commented
somewhere in an overridable file that will land in user space.

In a later PR, we can decide what the interface will be.