Add public cloud image prep script
Artifacted builds need to be done on a static starting
point in terms of the packages installed on the base OS
of the host used for the deployment.

If a static image is not used, then when the host image
is updated beyond the packages available in the artifacted
Apt repository, deployments will fail.

In Rackspace public cloud the base images are updated
frequently, so this script is used to take an exiting
public cloud instance and prep it to be used as a saved
image.

In practise the image used for testing can be updated
in the same cadence as the releases. However it must be
clear which version of the image has been used to test
a particular release tag to ensure that a build is always
repeatable.