Remove unused rpcd group_vars
In 35a375c7ff8104578ee93bfd6b1d9d3b612b2e62 a
new set of group_vars was implemented with the
vars plugin that merges both the RPC-O and OSA
group vars.

As such, none of these group_vars are used or
necessary any more.

The ansible-role-requirement for the
rcbops_openstack-ops role has been updated to
include a fix caught when testing this patch.