Fix python artifact build for ceilometer/octavia
The py_pkgs lookup does not look into group_vars
for repo information, causing the defaults from
the role to be used.

This patch moves the octavia repo details back
to the overrides file to ensure that the CI repo
build for artifacting is able to succeed.

To pick up the require group_vars for RPC,
including the variable
repo_build_upper_constraints_overrides
the artifact build scripts are adjusted to
ensure that the openstack-ansible.rc file is
appropriately adjusted to consume them.