Move logstash role out and add log aggregation
This change is for issue/942 with the overall goal of filtering
and pushing customer logs from the local elk stack to a shared
global elk cluster. This should give support easier access to
search customer logs for a quicker resolution with customer issues.
The existing logstash role was moved out to rcbops/rpc-role-logstash
with additions to push logs to a central elasticsearch cluster.
Enabling this should be a simple OSA overrides update and a re-run
of the logstash playbook.

Connects rcbops/u-suk-dev#942