UG-534 Check if router moves agent during upgrade
This commit adds some tasks to the pre and post upgrade roles to check
if a neutron router has moved agents during the upgrade.  If a move has
been detected, a debug message is output so the deployer can
investigate further.

Note that we save the neutron router agent to a pre and post file in
/tmp, and then compare the two.  The post upgrade files don't
necessarily need to be written to disk, but perhaps there is value
leaving these files around so the deployer can manually inspect.  At
present they will get removed, however this commit can be updated to
leave them in place or to write them to a more suitable location.

(cherry picked from commit 4a4fa95fd2f6491838e8f1cf6924c897c01f324c)