Adapt apt repo use based on artifact availability
In 698390d69b9398d631275ee57aec6cdecbe4b00e the
process was made to be more adaptable, but the
default config to setup the repo was missed.

This patch closes the gap.

Note that this will destroy every line in the
defaults file to the end of the file, so that
further patches which implement configuration
that matters must be above the line starting
with 'rpco_mirror_base_url'.