Save neutron router output to a file
This commit mirrors the backup directory vars in rpc_pre_upgrade in
rpc_post_upgrade so that we can use the same backup dir from each.
We also glue these vars together by defining them in group_vars.

Additionally, we modify rpc_post_upgrade to slurp the contents of
the pre-upgrade files so we can compare the actual content, not just
checksum.  This allows us to see if a router no longer has an agent
assigned.

Lastly, we render a template (neutron-router-status.txt.j2) in
the post upgrade role which succinctly informs a deployer if there are
any routers that need investigation.