Remove support role from standard testing
As the role's purpose is to implement access and
convenience scripts for operational purposes, it
really does not fit well in development testing.

The role takes over 20 minutes to execute on an
AIO due to the way it is implemented [1]. It is
also completely untested [2].

It should rather be implemented as some sort of
late stage testing before release, rather than
waste time and resources for PR testing.

[1] https://github.com/rsoprivatecloud/openstack-ops/issues/33
[2] https://github.com/rsoprivatecloud/openstack-ops/issues/20