Remove neutron checksum workaround
The checksum addition using this method was very
unreliable, so in OSA it was added [1] in a more
robust way to the neutron role and removed from
run-playbooks [2].

RPC-O therefore inherits it and does not need
to do this any more. In fact it may interfere
with it so it's better not to use this hack.

[1] https://review.openstack.org/330405
[2] https://review.openstack.org/330389