Add TODO for patcher removal
The container build process currently relies on
removing part of the roles in order to prevent
ssh keys being implemented in the artifacts.

This patch adds a note to reflect the upstream
patches which will remove the need for this.

(cherry picked from commit 4c9ce891ae2793f9316ebc00e914f57f438657d8)