AR-199 Seperate apt artifacts per series
Currently the snapshot merges are only based on the
distro used. This does not work across series when
the snapshots for the same distro need different
source snapshot sets (eg MariaDB 10.0 vs 10.1).

This sets up the snapshot map so that we can map
them properly to the series concerned.

In order to facilitate looping through the snapshot
sets that need to be created and published, some
playbook re-organisation was required. The same was
equired for the apt artifact testing.

A little more whitespace was also added between
plays and tasks to make them easier to read.

Conflicts:
scripts/artifacts-building/apt/aptly-vars.yml

Backport notes:
Removed r15 from the aptly vars as it is not
applicable to this branch

(cherry picked from commit 5cb495fecf4a47dcb0b1e2965e6871b8cb9bd187)