AR-106 Consolidate mirror sync into playbook
In order to make the aptly-all playbook easier
to follow and read, the sync-with-mirror play
is consolidated into it and the abstraction
of 'actions' is removed.

For the pull from rpc-repo, the synchronize
task's 'recursive' option is removed as it is
set to the default value for the module.

For the push to rpc-repo the synchronize
task's 'delete' option is enabled to ensure
that any changes made to the aptly database
are pushed up to rpc-repo. Without this the
process of cleaning has to be re-done over
and over again, and rpc-repo will retain unused
packages.

(cherry picked from commit de35320702d732017deedcac572c16e3a2f7140f)