Make pre_redeploy script friendlier (#2376)
When running outside gates, the pre_redeploy script didn't
ask for a verification of the variable migration.

This commit ensures to ask a user verification of the variable
migration, but also notices the user shouldn't interrupt the
process to verify them. As a result, the process should be
more user-friendly.