Don't override maas_auth_url
maas_auth_url is already defined in rpc-maas in maas agent vars.
Copying auth_url into a user_* variable file will override this.
We shouldn't force this behavior.

On top of this, the value used is undefined.
If we still want to keep an user override named
rackspace_cloud_auth_url, we should at least default it to a
safe value.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>