LA-368 Remove lxc_container_cache_files from vars
`lxc_container_cache_files` is a list of files to copy from the
deployment host. The file, `"/etc/apt/sources.list.d/rcpo.list"`, is
created by the role `pip_install` on the required hosts and so there
should not be anything to copy. In general, having this variable set
does not cause an issue because the first infra node is used as the
deployment node. Where the deployment node is not itself part of the
deployment, e.g. on a multi-node AIO, the file is missing and causes a
failure in the `lxc_hosts` role when performing a leapfrog upgrade.