LA-365 Force yaprt to build setuptools 24
There is no constraint at all for setuptools in the rpc builds.
So we can install any version of these software, and yaprt
is building latest wheels.

We need to not use latest wheels for setuptools, therefore
I list it as a pip_package. It will therefore be listed in the
deps, and will be used in the yaprt process. No new setuptools
will be used, as the dependency will be met.

On top of it, it was free for all for elasticsearch, so I took
the liberty (pun intended) to strengthen the elasticsearch and
elastisearch-curator constraints, with a value coming from
Liberty-12.2 branch.

Also constrains beaver dependencies.