K8S-116 Remove magnum deployment
The approach for implementing containers is
going to be completely re-done. This patch
removes the current approach to prevent bit
rot.

The remaining Octavia bits are adjusted to
ensure that the SHA's are fixed at a point
in time. As the development work for that
is progressed, those SHA's may be adjusted.