ISU-103 Update OSA SHA and rpc_release
We want to be able to release based on the
latest tested artifacts, so it makes better
sense to use the next release's version in
rpc_release. We can still implement rc tags
if we want to, but the rpc_release will not
reflect the tag.

This sha bump includes the rolling-upgrade work
which has been implemented in OSA in order to
support intra-series upgrades.