RE-155 Add rpco apt source to lxc base cache
In order for the lxc cache prep process to make use
of the extra rpco apt source we need to ensure that
the repo source file is copied into the cache prior
to doing any package actions. This patch ensures
that the repo config is there.

The OSA SHA is also updated to ensure that the
unbound variable bug is fixed when installing pip.
ref: https://review.openstack.org/489260